requirements.txt
django-annoying
django-extensions
django-uni-form
